# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clean Architecture Study Api",
        "description": "クリーンアーキテクチャー学習用のAPIです。",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "User/Auth"
                ],
                "description": "ユーザー新規登録",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "氏名",
                                        "type": "string",
                                        "default": "Ippei Kamimura"
                                    },
                                    "email": {
                                        "description": "メールアドレス",
                                        "type": "string",
                                        "default": "ippei_kamimura@icloud.com"
                                    },
                                    "password": {
                                        "description": "パスワード",
                                        "type": "string",
                                        "default": "aaaaaa"
                                    },
                                    "password_confirmation": {
                                        "description": "パスワード(確認)",
                                        "type": "string",
                                        "default": "aaaaaa"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "認証トークンを返す"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "User/Auth"
                ],
                "description": "ログイン",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "メールアドレス",
                                        "type": "string",
                                        "default": "ipp@icloud.com"
                                    },
                                    "password": {
                                        "description": "パスワード",
                                        "type": "string",
                                        "default": "aaaaaaaa"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "認証トークンを返す"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "User/Auth"
                ],
                "description": "自分自身を取得",
                "operationId": "b9abb1a7a74670a19c215c2c133f14d8",
                "responses": {
                    "200": {
                        "description": "自分自身のJson"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/payment/account": {
            "put": {
                "tags": [
                    "User/Payment"
                ],
                "description": "pamentのアカウントの更新",
                "operationId": "3ddc2d74d90e188b60562eb8c08fa975",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "description": "クレカのトークン",
                                        "type": "string",
                                        "default": "token"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User/Payment"
                ],
                "description": "pamentのアカウントを作成",
                "operationId": "c3e7517b2eafdee06b1d5d53eb1417ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "description": "クレカのトークン",
                                        "type": "string",
                                        "default": "token"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/product": {
            "post": {
                "tags": [
                    "User/Product"
                ],
                "description": "Productを作成",
                "operationId": "1d0b0d8724685b7b8f9d87f403bfc4df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Product名",
                                        "type": "string",
                                        "default": "花びら大回転"
                                    },
                                    "price": {
                                        "description": "価格",
                                        "type": "integer",
                                        "default": "0"
                                    },
                                    "stock": {
                                        "description": "在庫",
                                        "type": "integer",
                                        "default": "0"
                                    },
                                    "shop_id": {
                                        "description": "shopのID",
                                        "type": "string",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "作成したShopのオブジェクト"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/shop": {
            "post": {
                "tags": [
                    "User/Shop"
                ],
                "description": "Shopを作成",
                "operationId": "c6201d49b307315866578fdcf3e08b8c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Shop名",
                                        "type": "string",
                                        "default": "ビンコムセンター"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "作成したShopのオブジェクト"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}